plot_stats.py plotting of Ob impact made look better.
ObImpact.py
Obs. Impact is now calculated independently of the type of model. All it needs is the assimilation cycles.
This program will calculate Ob. Impact and store it in a binary data file.

proc_ObImpact.py
Reads the binary data file from ObImpact.py and processes it for plotting etc.