src/ObImpact.py Previously, scalar (or total) ???_dJ, ???_dJa and ???_dJb were saved, where ??? == either ens or adj. Now onwards, ???_dJa and ???_dJb will be saved, where in these are for each observation.
Thus OLD ???_dJa = NEW np.nansum(???_dJa) and same for dJb.

???_dJ is simply obtained as np.nansum(???_dJa) + np.nansum(???_dJb)

The advantage of saving for every observation is one can get the observation impact for each
observation and not just the total observation impact which is the sum of impacts for all
observations.

proc_ObImpact.py will need to be modified.