NOTE -- a pre-mature commit was performed at r145 before the entire log was written. A minor tweak is done in this commit tht does not change anything.
A LXX_stats.py -- in r145
D L63_stats.py -- in r145
D L96_stats.py -- in r145

in r145, LXX_stats.py replaces L63_stats.py and L96_stats.py for generic-city.
LXX_stats.py is not tested. If in the future, this is used, it would be WISE to test this script
with L96_stats.py