M    ensvarDA.py M    module_DA.py M    L96_param_ensvarDA.py
a new function check_ensvarDA is added in module_DA that checks varDA and ensDA options.
it also makes sure that varDA is using the sqrtB preconditioning as this is the only way currently
implemented.
If one wants to use no preconditioning or full B preconditioning, hybDA may be suitable with static
contribution as zero.